feat: limit attachment size
