+++ packages/blocks/src/__internal__/utils/filesys.ts
@@ -223,11 +223,17 @@ export const uploadImageFromLocal = async (storage: BlobManager) => {
   return res;
 };
 
-export const uploadFileFromLocal = async (storage: BlobManager) => {
+export const uploadFileFromLocal = async (
+  storage: BlobManager,
+  beforeUpload?: (file: File) => boolean | Promise<boolean>
+) => {
   const file = await openFileOrFiles({
     acceptType: 'Any',
   });
   if (!file) return;
+  const allowUpload = beforeUpload ? await beforeUpload(file) : true;
+  if (!allowUpload) return;
+
   // The original file name can not be modified after the file is uploaded to the storage,
   // so we create a new file with a fixed name to prevent privacy leaks.
   const anonymousFile = new File([file.slice(0, file.size)], 'anonymous', {
